feat(dashboard-activity): Handle logic for showing badge on Favourite Menu
